Отрисуй меня полностью (часть 1) исправления
